configure: Fix summary shown standard when passing -c++std c++14
When -c++std c++14 is passed to configure, we call
 qtConfCommandlineSetInput("c++14", "yes")
 qtConfCommandlineSetInput("c++1z", "no")

But the real feature is called 'c++17'. 'c++1z' is also a feature, but
not the main one, it is derived from the 'c++17' feature.

The 'c++17' has the following condition
 "features.c++14 && tests.cxx17",

The condition would evaluate to true, setting config.input.c++1z to
'no' would not override the "enable" / "disable" value
and thus the feature summary would pick up c++17 as the first
available feature standard.

Fix this by setting the feature input of both "c++17" and "c++1z".

Fixes: QTBUG-124179
Change-Id: Id548834cc1dd2652ae699c9995c72e163a00f41d
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>