Android: account for namespace in build.gradle instead of manifest
AGP version 7.4 deprecated the use of "package" attribute in the
manifest to specify the unique package name, it's instead been
moved to build.gradle file and set using "namespace" property.
This patch adds support of that to androiddeployqt.

Removing the "package" attribute from the default manifest would
break Qt Creator 13 and below because Qt Creator would fail to
deploy apps without such attribute in the manifest. For that reason
we'll defer removing it until a later version, for example Qt 6.10,
to allow some buffer for a Qt Creator that can handle that to be
adopted by users to reduce breakage.

[ChangeLog][Android] Add support for namespace in build.gradle instead
of the package attribute in the manifest.

Task-number: QTBUG-128362
(cherry picked from commit d32f7dc0165489e96dd5e2834b513b289b46d185)
Change-Id: I6940ce58f33f844715f46ed00a6ff778d1f43770
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>