QKdeTheme: update color scheme with unknown desktop theme
When a KDE theme name doesn't contain the keywords "dark" or "light",
QKdeThemePrivate::updateColorScheme() falls back to heuristics. It
compares text and base color. It sets the color scheme to dark, if
text is lighter than base and otherwise to light.

When the theme name can't be read from KDE settings, because it is
stored on a different location or not yet set, the heuristics fallback
doesn't happen. The color scheme always becomes "Unknown".

Remove this limitation to the benefit of currently unsupported Linuxes.

Fixes: QTBUG-125285
Change-Id: I120dd23eee71b14ea8ce3735c3f53bda2caa403e
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 71d0de98359decfcc9be01a0ed5823a9595ef5d4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>