QIcon: Adjust documentation of QIcon::pixmap()
The documentation of QIcon::pixmap() taking a devicePixelRatio could be
read in the way that a QPixmap of the requested size and
devicePixelRatio is generated which is not true. The generation part is
only valid for the given mode and state.

Pick-to: 6.5
Change-Id: I5374cd7be17c32cd87d7b35552e60a4c07ebe6b2
Reviewed-by: Samuel Gaist <samuel.gaist@idiap.ch>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit 2422f8b18dd325e5ac7d0adff39b2c93a8f75466)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>