QByteArray: deduplicate QtPrivate::findByteArray() and qFindByteArray()
The former is the only place where the latter is called and both were
doing duplicate and redudant work. So merge the two functions again...
and split off the actual hashed searching into a new qFindByteArray.
That simplifies the diff and Git history.

Drive-by adding the Q_NEVER_INLINE to the Boyer-Moore search, which will
be reached via tail-call.

Task-number: QTBUG-125283
Change-Id: If05cb740b64f42eba21efffd17cf2ed2bbc64b17
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>