Add support for Arm64EC
Arm64EC (Emulation compatible) is Microsoft's solution for transitioning
from x64 to Arm64
(https://learn.microsoft.com/en-us/windows/arm/arm64ec).

In this configuration a user can mix and match binaries and libraries
targetting Arm64EC and x64 and run the result on a Windows on Arm
machine.

In order to target ARM64EC the following requirements have to be met:
- Usage of Visual Studio 2026
- An Arm64 development environment for MSVC
- Addition of /arm64EC to CFLAGS
- Addition of /arm64EC and /d2arm64ECMarkAllFuncsPatchable to CXXFLAGS

As most details happen "behind the scenes" in Visual Studio it is hard
to reflect the changes in cmake properly. For now the CFLAGS and
CXXFLAGS have to be added manually. Maybe it makes sense to set them
accordingly when a corresponding target platform is set.

The change is cherry picked to 6.11 so that we can get more feedback by
having experimental support for this new configuration.

Task-number: QTBUG-125260
Change-Id: Ia2404fdcdbd3c0d443c25649af84aeb0486fcbcd
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit f7d7fb8c6f710fa094ebeb7b28c7e99d974cd2da)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>