Doc: Hint about QtMsgType order
You could get the idea that the underlying int value is indicative
of the severity. Make it more explicit that the order in the text
is the correct one, and reorder for Qt 7.

While at it, remove mentioning of Qt 5.5.

Fixes: QTBUG-102347
Change-Id: I5d1b46afb37e50baa209a06c192a943980ca774b
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit 469367dc08654e02dae39162750fce7fa39d906d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>