QCryptographicHash: extend hashInto to more than one piece of data
This allows use of the noexcept static function in cases where more
than one piece of data needs to be hashed, and concatenation of said
data would have to allocate memory.

Port QUuid to use the new function, allowing to mark the V3 and V5
create functions noexcept.

As a drive-by, take QUuid by value in the internal helper function,
fixing a Clazy warning.

Task-number: QTBUG-125431
Change-Id: I17938f0be44c91085e2aaa5574953f8dceacc990
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>