coin: Add instructions for generating an SBOM when installing Qt
The instructions will only run if a coin platform config has the
'GenerateSBOM' feature.

The -sbom configure option is not sticky across repos,
similar to -headersclean, so that if a user wants to build another
qt repo against a prebuilt qt, it won't try to generate the sbom
automatically.
That's why we append -sbom to all the repos configure args, and not
just the qtbase one.

Pick-to: 6.8
Task-number: QTBUG-122899
Change-Id: Ice91cbc430826838119ce3085cbd52c0909f2746
Reviewed-by:  Alexey Edelev <alexey.edelev@qt.io>