Add licenses present in SPDX tags or REUSE.toml file to LICENSES dir
The licenses are present in 3rd party directories. Those licenses are now
present in some SPDX tags or in REUSE.toml file. qtqa license test and
reuse need them to be in the LICENSES directory.
If the file with the license text is present in the 3rd party directory
a symlink is created. The license text is added as a new file when the
license present in the 3rd party is grouping more than one license text
in a given file.

Task-number: QTBUG-125569
Change-Id: I3e2471b435e83b2fcd024a4938d193a14712a460
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>