tst_QSettings: make it clear testErrorHandling isn't run
That is, make it a QSKIP everywhere instead of a silent PASS on Unix
systems.

This test was accidentally disabled in Qt 4.6 commit
31e358f2290c145b839fc5b7b277922c1ab6e19b ("fix tests for QT_NO_PROCESS
and when running tests as root"), which introduced an #else that
disabled the entire meat of the test (that is, the test was for
everything that is neither Unix nor Windows, which is an empty set).

I am not enabling this test right now because it's failing and I don't
plan on spending the time to fix it. It is possible to narrow down the
QSKIP to QEXPECT_FAIL on the actual failing test lines, but that is also
left as an exercise to the reader.

FAIL!  : tst_QSettings::testErrorHandling(0600 0100) Compared values are not the same
   Actual   ((int)settings.status()): 1
   Expected (statusAfterSetAndSync) : 0
   Loc: [tests/auto/corelib/io/qsettings/tst_qsettings.cpp(941)]
FAIL!  : tst_QSettings::testErrorHandling(0600 0100) 'settingsDir.removeRecursively()' returned FALSE. ()
   Loc: [tests/auto/corelib/io/qsettings/tst_qsettings.cpp(314)]
FAIL!  : tst_QSettings::testErrorHandling(  -1 0300) 'settingsDir.removeRecursively()' returned FALSE. ()
   Loc: [tests/auto/corelib/io/qsettings/tst_qsettings.cpp(314)]
FAIL!  : tst_QSettings::testErrorHandling(0600 0300) 'settingsDir.removeRecursively()' returned FALSE. ()
   Loc: [tests/auto/corelib/io/qsettings/tst_qsettings.cpp(314)]
FAIL!  : tst_QSettings::testErrorHandling(0600 0500) Compared values are not the same
   Actual   ((int)settings.status()): 1
   Expected (statusAfterSetAndSync) : 0
   Loc: [tests/auto/corelib/io/qsettings/tst_qsettings.cpp(941)]
FAIL!  : tst_QSettings::testErrorHandling(0600 0500) 'settingsDir.removeRecursively()' returned FALSE. ()
   Loc: [tests/auto/corelib/io/qsettings/tst_qsettings.cpp(314)]

Task-number: QTBUG-126008
Change-Id: If3345151ddf84c43a4f1fffd17d277979e987905
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>