QByteArray: use qulltoString_helper() to make strings
Mårten made this code ready for QByteArray in commits
e5e8e4f59ba4c3d59303220a57677360992eb553 and
5d0542a356be1c47e2f99e6ba38b8e520ca0710c, but never actually made
QByteArray use it.

The simplest change of moving qulltoa2() is sufficient, instead of
moving the template to the header and using it in qbytearray.cpp.
qulltoa2() is called in three places and we don't want the compiler to
inline the function there, so having the function in a separate TU
accomplishes that for us. GCC won't inline it in LTO builds either.

Change-Id: If3345151ddf84c43a4f1fffd17d28682445fbdaa
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>